@@ -24,7 +24,7 @@
     "@angular/platform-browser": "6.0.0-beta.8",
     "@angular/platform-browser-dynamic": "6.0.0-beta.8",
     "@angular/router": "6.0.0-beta.8",
-    "@ngxs/store": "^2.0.0-rc.23",
+    "@ngxs/store": "^2.0.0",
     "angularfire2": "^5.0.0-rc.6",
     "core-js": "^2.4.1",
     "firebase": "^4.12.1-0",
@@ -50,6 +50,6 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.0",
     "tslint": "~5.9.1",
-    "typescript": "2.7.2"
+    "typescript": "^2.8.1"
   }
 }
