@@ -6,6 +6,5 @@ module.exports = {
   singleQuote: true,
   trailingComma: 'none', // other options `es5` or `all`
   bracketSpacing: true,
-  arrowParens: 'avoid', // other option 'always'
-  parser: 'typescript'
+  arrowParens: 'avoid' // other option 'always'
 };
