@@ -1,9 +1,11 @@
+import { first } from 'rxjs/operators';
+
 import { Injectable } from '@angular/core';
 
+import { Actions, ofAction } from '@ngxs/store';
+
 import { AuthStateService } from '../../core/services/state/auth-state.service';
-import {Actions, ofAction} from "@ngxs/store";
-import {UpdateUser} from "../../shared/actions/auth.action";
-import {first} from "rxjs/operators";
+import { UpdateUser } from '../../shared/actions/user.action';
 
 @Injectable()
 export class AppLoadService {
@@ -11,6 +13,6 @@ export class AppLoadService {
 
   async initAuth(): Promise<void> {
     await this.as.initAuth();
-    await this.actions$.pipe(ofAction(UpdateUser), first()).toPromise()
+    await this.actions$.pipe(ofAction(UpdateUser), first()).toPromise();
   }
 }
