@@ -1,17 +1,16 @@
-import { User, Persistence } from '@firebase/auth-types';
+import { Persistence } from '@firebase/auth-types';
 
 import { LoginProvider } from '../enums/login-provider.enum';
+import { AuthPhase } from '../enums/auth-phase.enum';
 
 export class LoginWithProvider {
   constructor(readonly provider: LoginProvider) {}
 }
 
-export class LogoutUser {}
-
 export class SetPersistence {
   constructor(readonly persistence: Persistence) {}
 }
 
-export class UpdateUser {
-  constructor(readonly user: User | null) {}
+export class SetPhase {
+  constructor(readonly phase: AuthPhase) {}
 }
