@@ -1,6 +1,8 @@
-export interface IUser {
-  email: string | null;
-  displayName: string | null;
-  photoURL: string | null;
-  isNewUser: boolean;
+import { IData } from './data.interface';
+
+export interface IUser extends IData {
+  email?: string | null;
+  displayName?: string | null;
+  photoURL?: string | null;
+  isNewUser?: boolean;
 }
