@@ -3,14 +3,12 @@ import * as firebase from 'firebase/app';
 import { Action, Selector, State, StateContext } from '@ngxs/store';
 import { AngularFireAuth } from 'angularfire2/auth';
 
-import { IUser } from '../interface/user.interface';
-import { LoginWithProvider, LogoutUser, SetPersistence, UpdateUser } from '../actions/auth.action';
+import { LoginWithProvider, SetPersistence, SetPhase } from '../actions/auth.action';
 import { LoginProvider } from '../enums/login-provider.enum';
 import { AuthPhase } from '../enums/auth-phase.enum';
 import { HideLogin } from '../actions/ui.action';
 
 export interface AuthStateModel {
-  user: IUser | null;
   phase: AuthPhase;
   provider?: LoginProvider;
   persistence?: firebase.auth.Auth.Persistence;
@@ -19,8 +17,7 @@ export interface AuthStateModel {
 @State<AuthStateModel>({
   name: 'auth',
   defaults: {
-    phase: AuthPhase.LoggedOut,
-    user: null
+    phase: AuthPhase.LoggedOut
   }
 })
 export class AuthState<T extends StateContext<AuthStateModel>> {
@@ -29,12 +26,12 @@ export class AuthState<T extends StateContext<AuthStateModel>> {
     [LoginProvider.Facebook]: firebase.auth.FacebookAuthProvider
   };
 
-  @Selector()
-  static user(state: AuthStateModel): IUser | null {
-    return state.user;
-  }
-
   constructor(readonly afAuth: AngularFireAuth) {}
+
+  @Selector()
+  static loggedIn(state: AuthStateModel): boolean {
+    return state.phase === AuthPhase.LoggedIn;
+  }
 
   @Action(SetPersistence)
   async setPersistence({ patchState }: T, { persistence }: SetPersistence): Promise<void> {
@@ -45,36 +42,16 @@ export class AuthState<T extends StateContext<AuthStateModel>> {
 
   @Action(LoginWithProvider)
   async loginWithProvider({ patchState, dispatch }: T, { provider }: LoginWithProvider): Promise<void> {
-    patchState({ provider, phase: AuthPhase.Authenticating });
+    patchState({ provider });
+    dispatch(new SetPhase(AuthPhase.Authenticating));
 
     await this.afAuth.auth.signInWithPopup(new this.authProviders[provider]());
 
     dispatch(new HideLogin());
   }
 
-  @Action(LogoutUser)
-  async logoutUser({ patchState }: T): Promise<void> {
-    patchState({ provider: undefined, phase: AuthPhase.LoggedOut, user: null });
-
-    await this.afAuth.auth.signOut();
-  }
-
-  @Action(UpdateUser)
-  updateUser({ patchState }: T, { user }: UpdateUser): void {
-    let phase: AuthPhase = AuthPhase.LoggedOut;
-    let newUser: IUser | undefined;
-
-    console.log(user);
-    if (user) {
-      phase = AuthPhase.LoggedIn;
-      newUser = {
-        email: user.email,
-        displayName: user.displayName,
-        photoURL: user.photoURL,
-        isNewUser: user.metadata.lastSignInTime === user.metadata.creationTime
-      };
-    }
-
-    patchState({ user: newUser, phase });
+  @Action(SetPhase)
+  setPhase({ patchState }: T, { phase }: SetPhase): void {
+    patchState({ phase });
   }
 }
